Fix #7: "Edit entry" creates duplicates
The issue was mainly an off-by-one error: when extracting the content of the chapter to be replaced, we were adding the first line of the next chapter, which created a duplicate.

I also removed the "reordering" of the edited chapter, as I believe the user expects the chapter to stay where it is after being edited.