Feature: add support for "generic flags"
Currently the UI handles only three types of flags: final, death and fixed; an user could technically add more by adding the appropriate tags to the chapter body, but the UI would not handle them:
- they would not be shown anywhere
- they would be removed if the "edit panel" is used

The latter is a problem, as the user should be able to freely use the "edit panel" without fear of losing information.

With this commit we support generic flags:
- they are shown in the sidebar with a custom image
- they can be edited in the "edit panel" through a textarea
- they are used to colour in a different way the nodes in the grap viz

I also took the chance and fixed a bug when multiple flags were incorrectly reported on one line when using the "edit panel".